GP_SERV_COMMAND_BAZAAR_SHOPPING
