+++ src/map/packets/c2s/0x105_bazaar_list.cpp
@@ -22,7 +22,7 @@
 #include "0x105_bazaar_list.h"
 
 #include "entities/charentity.h"
-#include "packets/bazaar_check.h"
+#include "packets/s2c/0x108_bazaar_shopping.h"
 #include "packets/bazaar_item.h"
 
 auto GP_CLI_COMMAND_BAZAAR_LIST::validate(MapSession* PSession, const CCharEntity* PChar) const -> PacketValidationResult
@@ -45,7 +45,7 @@ void GP_CLI_COMMAND_BAZAAR_LIST::process(MapSession* PSession, CCharEntity* PCha
 
         if (!PChar->m_isGMHidden || (PChar->m_isGMHidden && PTarget->m_GMlevel >= PChar->m_GMlevel))
         {
-            PTarget->pushPacket<CBazaarCheckPacket>(PChar, BAZAAR_ENTER);
+            PTarget->pushPacket<GP_SERV_COMMAND_BAZAAR_SHOPPING>(PChar, GP_SERV_COMMAND_BAZAAR_SHOPPING_STATE::Enter);
         }
 
         PTarget->BazaarCustomers.emplace_back(EntityID);
